Adding graph request test helper
